==> 01_phen_phth_conc.Rmd <==
---
title: "01_phen_phth_concen"
author: "Puvvula"
date: "2023-06-23"
output: pdf_document
---

#chemical concentrations
```{r}
library(pacman)
pacman::p_load(tidyverse, janitor)
data <- "~/Documents/phth_phe_MWAS/data/"
result<- "~/Documents/phth_phe_MWAS/result/"
analy_dat<- "~/Documents/phth_phe_MWAS/analy_dat/"
```

#extracting maternal exposure data
```{r}
#get csv file names contain phenols and phthalates
file_list <- list.files(path = data,
                        pattern = "^report.*\\.csv",
                        full.names = TRUE,
                        recursive = TRUE)

#convert all the variable names to lower letter case and charecter variables to factors
df_list <- lapply(file_list, function(file) {
  df <- read_csv(file)
  df <- df |> mutate_if(is.character, as.factor)
  df <- clean_names(df)  # Convert variable names to lowercase and clean them
  return(df)
})

#bind phenols and phthalate observations
dat <- bind_rows(df_list)

#imputing values <LOD with LOD/sqrt-2
dat <- dat |> mutate(lod = as.character(lod), # Convert lod to character
         lod = parse_number(lod), # Extract numeric value from "lod" variable
         level = ifelse(!is.na(lod), round(lod/sqrt(2),2), level))

#summary table
summary_stats <- dat |>
  group_by(chemical) |>
  summarise(total_observations = n(),
            pct_below_LOD = round((sum(!is.na(lod))/total_observations)*100, 2), 
            median_level = median(level, na.rm = TRUE), # Calculate median of "level"
            percentile_25 = quantile(level, 0.25, na.rm = TRUE), # Calculate 25th percentile of "level"
            percentile_75 = quantile(level, 0.75, na.rm = TRUE)) # Calculate 75th percentile of "level"

#export raw summary pre imputation
write_csv(summary_stats, paste0(result, "raw_summary.csv"))
```

#standardized and creatinine adjusted values
#cr_adj = (urine_conc / urine_creat) * 100 (unit ng/g creatinine)
```{r}
#reading urine creatinine
ur_creat<- read_csv(paste0(data, "phenols/Creatinine_report 2019-0002-01.csv")) |>
  clean_names() |>
  rename(participant_id = "external_id_number_1", ur_crt= "reportable_result")

#calculated creatinine adjusted chemical concentrations
dat_exp <- dat |>
  mutate(level_crt_adj = round(level / ur_creat$ur_crt[match(participant_id, ur_creat$participant_id)]*100, 2)) |>
  select(participant_id, chem_short, level_crt_adj) |>
  pivot_wider(names_from = chem_short, values_from = level_crt_adj) |>
  clean_names() |>
  rename(group="participant_id")

# Create a summary table for variables in wide format
urine_crt_adj_chem_summary <- dat_exp |>
  gather(key = "variable", value = "value", -1) |>
  group_by(variable) |>
  summarize(median_level = round(median(value, na.rm = TRUE), 2),
            percentile_25 = round(quantile(value, 0.25, na.rm = TRUE), 2),
            percentile_75 = round(quantile(value, 0.75, na.rm = TRUE), 2))

#export processed exposure data
write_csv(dat_exp, paste0(analy_dat, "phen_phth_exp.csv"))

#export processed summary
write_csv(urine_crt_adj_chem_summary, paste0(result, "processed_summary.csv"))
```

================================================================================
#read questionniare data
```{r}
qn_dat<- read_csv("~/Documents/MWAS_home/pj_dat/pah_fin_impu.csv") |>
  select(c(1, 11:26)) |>
  mutate(stove = replace(stove, stove %in% c("NA_NA", "NA_lt_4hr"), NA))|>
  mutate(stove = case_when(
    grepl("Gas", stove, ignore.case = TRUE) ~ "Gas",
    grepl("Electric", stove, ignore.case = TRUE) ~ "Electric",
    is.na(stove) ~ NA_character_,TRUE ~ stove)) |>
  mutate(hh_inc_cat = if_else(is.na(hh_inc_cat), names(which.max(table(hh_inc_cat))), hh_inc_cat)) |>
  mutate(stove = if_else(is.na(stove), names(which.max(table(stove))), stove)) |>
  mutate(bbquse = if_else(is.na(bbquse), names(which.max(table(bbquse))), bbquse)) |>
  mutate(dist_majhw_cat = if_else(is.na(dist_majhw_cat), names(which.max(table(dist_majhw_cat))), dist_majhw_cat)) |>
  mutate_if(is.character, as.factor) |>
  rename(barbeque_use="bbquse", distance_to_highway="dist_majhw_cat", education="mom_edu_cat", income= "hh_inc_cat",
         stove_use="stove", race="race_cat")
```
================================================================================

#join questionniare data - filtered data that contain either mom or baby metabolome data
```{r}
dat_fin <- left_join(dat_exp, qn_dat, by ="group") |>
  filter(!(is.na(participant_id_mom) & is.na(participant_id_baby)))

#export final analysis dataset - use this data for any further analysis !
#data need to be log transformed
write_csv(dat_fin, paste0(analy_dat, "phen_phth_fin.csv"))
```
