==> 04_enrichment_figs.Rmd <==
---
title: "04_enrichment_figs"
author: "Puvvula"
date: "2024-01-08"
output: pdf_document
---

#read merged pathway analysis results
```{r}
dat <- read_csv("~/Documents/phth_phe_MWAS/result/enrich_results_combined_march.csv") |>
  clean_names() |>
  rename(p_gamma = "gamma", 
         pathway = "x") |>
  filter(p_gamma < 0.05) |>
  mutate(logp = -log10(p_gamma), enrichment = hits_sig/expected) |>
  mutate(enrichment_disc = cut(enrichment, breaks = seq(0, 18, by = 2), 
                                 labels = c("< 2", "2 - <4", "4 - <6",
                                            "6 - <8", "8 - <10", "10 - <12", "12 - <14", "14 - <16", "16 - <17"))) |>
  filter(pathway_total > 3)  |>
  mutate(chemical = fct_recode(chemical, 
                               "BPA" = "BP-A", 
                               "BPS" = "BP-S",
                               "2,5-Dichlorophenol" = "2,5-DCP",
                               "2,4-Dichlorophenol" = "2,4-DCP", "Benzophenone-3" = "BP-3", "MCNP" = "mCNP")) |>
  filter((chemical == "BPA" & mom_baby %in% c("Maternal", "Newborn")) |
         (chemical == "BPS" & mom_baby %in% c("Maternal", "Newborn")) |
         (chemical == "2,5-Dichlorophenol" & mom_baby == "Newborn") |
         (chemical == "MBP" & mom_baby == "Newborn") |
         (chemical == "MiBP" & mom_baby == "Maternal") |
         (chemical == "MBzP" & mom_baby == "Newborn") |
           (chemical == "MECPTP" & mom_baby == "Maternal") | 
           (chemical == "MEHHTP" & mom_baby == "Maternal")
         )
```

#add KEGG supergroups to the pathway analysis results
```{r}
# Applying the function to your dataset
dat <- assignGroup(dat, 
                     pathway_column = "pathway", 
                     pathway_list = pathways)
```

#define sequence of the KEGG supergroups
```{r}
dat$group<- factor(dat$group, levels = c( "amino acid", "Carbohydrate",
                                          "Glycan", "Lipid",
                                          "Cofactors_vitamins",
                                          "Energy",
                                          "Nucleotide", "Other_secondary_metabolites", "Terpenoids_polyketides",
                                          "Xenobiotics", "NA"))
```

```{r}
dat_grp_check<- dat |> select(c(1,17)) |> distinct()
```


#pathway heatmap - Plot
```{r}
#choose either line 51 or 52
dat|>
  ggplot(aes(x = factor(chemical, levels = c("MBP", "MiBP", "MBzP", "MECPTP", "MEHHTP", "2,5-Dichlorophenol", "BPA", "BPS")), y = pathway)) + 
  #ggplot(aes(x = factor(chemical, levels = c("MEP", "MBP", "MiBP", "MBzP", "MONP", "MCOP", "MCNP", "MECPTP", "MEHHTP", "2,4-Dichlorophenol", "2,5-Dichlorophenol", "BPA", "BPS", "Benzophenone-3")), y = pathway)) + 
  geom_tile(aes(fill = factor(enrichment_disc)), color = "black") +
  scale_color_viridis(discrete = TRUE) +
  xlab("") +
  ylab("") +
  theme_bw() +
  theme(axis.text=element_text(size=10), 
        axis.title=element_text(size=10,face="bold"),
        strip.text = element_text(size=10),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position="bottom",
        legend.justification="right",
        legend.box="horizontal",
        legend.box.just="center",
        legend.margin=margin(t=0.1, r=0.1, b=2, l=0.1),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.major.y = element_blank(),
        axis.text.x = element_text(angle = 35, hjust = 1, vjust = 1, size = 10),
        axis.text.y = element_text(size = 10), 
        panel.spacing.x=unit(0.8, "lines"),
        panel.spacing.y=unit(0.02, "lines"),
        strip.text.y = element_blank()
        )+
  guides(fill = guide_legend(nrow = 1))+
  facet_grid(group~ mom_baby, scales = "free", switch = "y", space = "free")+
  labs(fill = "Enrichment factor")
```

```{r}
ggsave("~/Documents/phth_phe_MWAS/result/pathway_enrich.tiff", 
       width = 10, height = 18, dpi=300)
```

#Pathway overlap plots - data prep
```{r}
dat_olap <- dat |>
  unite(filename, chemical, mom_baby, sep = " ") |>
  dplyr::select(pathway, filename) |>
  distinct(pathway, filename, .keep_all=TRUE) |>
  unnest(cols = filename) |>
  mutate(GenreMember=1) |>
  pivot_wider(names_from = filename, values_from = GenreMember, values_fill = list(GenreMember = 0)) |>
  as.data.frame() |>
  dplyr::select(`2,5-Dichlorophenol Newborn`, 
                `BPA Newborn`, 
                `BPS Newborn`,
                `MBP Newborn`, 
                `MBzP Newborn`,
                `BPA Maternal`, 
                `BPS Maternal`, 
                `MiBP Maternal`,
                `MECPTP Maternal`,
                `MEHHTP Maternal`)
```

#overlap plots
```{r}
upset(dat_olap, 
      nintersects = 1500,
      keep.order = TRUE,
      nsets = 5,
      sets.bar.color = c("#0072B2", "#0072B2", "#0072B2", "#0072B2", "#0072B2", "#E69F00", "#E69F00", "#E69F00", "#E69F00", "#E69F00"),
      sets = c("2,5-Dichlorophenol Newborn", "BPA Newborn", "BPS Newborn", "MBP Newborn", "MBzP Newborn", "BPA Maternal", "BPS Maternal", "MiBP Maternal", "MECPTP Maternal", "MEHHTP Maternal"),
      set_size.show = TRUE,
      text.scale = 1.5, point.size = 4, line.size = 1,
      mainbar.y.max = 8,
      set_size.scale_max = 73,
      mainbar.y.label="No.of pathways \noverlapped across chemical biomarkers",
      sets.x.label="Total no.of \naltered pathways")
```

#=============================================================================#
#=============================================================================#
#Pathway overlap plots - data prep
```{r}
dat_olap_crazy <- dat |>
  unite(filename, chemical, mom_baby, group, sep = " ") |>
  dplyr::select(pathway, filename) |>
  distinct(pathway, filename, .keep_all=TRUE) |>
  unnest(cols = filename) |>
  mutate(GenreMember=1) |>
  pivot_wider(names_from = filename, values_from = GenreMember, values_fill = list(GenreMember = 0)) |>
  as.data.frame() |>
  dplyr::select(
    #pathway,
    # 9
    `2,5-Dichlorophenol Newborn Carbohydrate`,
    `2,5-Dichlorophenol Newborn amino acid`,
    `2,5-Dichlorophenol Newborn Lipid`,
    `2,5-Dichlorophenol Newborn Cofactors_vitamins`,
    `2,5-Dichlorophenol Newborn Glycan`,
    `2,5-Dichlorophenol Newborn Nucleotide`,
    `2,5-Dichlorophenol Newborn Energy`,
    `2,5-Dichlorophenol Newborn Xenobiotics`,
    `2,5-Dichlorophenol Newborn Other_secondary_metabolites`,
    # 10
    `BPA Newborn Carbohydrate`,
    `BPA Newborn amino acid`,
    `BPA Newborn Lipid`,
    `BPA Newborn Cofactors_vitamins`,
    `BPA Newborn Glycan`,
    `BPA Newborn Nucleotide`,
    `BPA Newborn Energy`,
    `BPA Newborn Xenobiotics`,
    `BPA Newborn Other_secondary_metabolites`,
    `BPA Newborn Terpenoids_polyketides`,
    # 9
    `BPS Newborn Carbohydrate`,
    `BPS Newborn amino acid`,
    `BPS Newborn Lipid`,
    `BPS Newborn Cofactors_vitamins`,
    `BPS Newborn Glycan`,
    `BPS Newborn Nucleotide`,
    `BPS Newborn Energy`,
    `BPS Newborn Xenobiotics`,
    `BPS Newborn Other_secondary_metabolites`,
    # 6
    `MBP Newborn Carbohydrate`,
    `MBP Newborn amino acid`,
    `MBP Newborn Lipid`,
    `MBP Newborn Cofactors_vitamins`,
    `MBP Newborn Xenobiotics`,
    `MBP Newborn Other_secondary_metabolites`,
    # 9
    `MBzP Newborn Carbohydrate`,
    `MBzP Newborn amino acid`,
    `MBzP Newborn Lipid`,
    `MBzP Newborn Cofactors_vitamins`,
    `MBzP Newborn Glycan`,
    `MBzP Newborn Nucleotide`,
    `MBzP Newborn Energy`,
    `MBzP Newborn Xenobiotics`,
    `MBzP Newborn Other_secondary_metabolites`,
    # 9
    `BPA Maternal Carbohydrate`,
    `BPA Maternal amino acid`,
    `BPA Maternal Lipid`,
    `BPA Maternal Cofactors_vitamins`,
    `BPA Maternal Glycan`,
    `BPA Maternal Nucleotide`,
    `BPA Maternal Energy`,
    `BPA Maternal Xenobiotics`,
    `BPA Maternal Other_secondary_metabolites`,
    # 7
    `BPS Maternal Carbohydrate`,
    `BPS Maternal amino acid`,
    `BPS Maternal Lipid`,
    `BPS Maternal Energy`,
    `BPS Maternal Nucleotide`,
    `BPS Maternal Cofactors_vitamins`,
    `BPS Maternal Other_secondary_metabolites`,
    # 9
    `MiBP Maternal Carbohydrate`,
    `MiBP Maternal amino acid`,
    `MiBP Maternal Lipid`,
    `MiBP Maternal Cofactors_vitamins`,
    `MiBP Maternal Glycan`,
    `MiBP Maternal Nucleotide`,
    `MiBP Maternal Xenobiotics`,
    `MiBP Maternal Other_secondary_metabolites`,
    #10
    `MECPTP Maternal Carbohydrate`,
    `MECPTP Maternal amino acid`,
    `MECPTP Maternal Lipid`,
    `MECPTP Maternal Cofactors_vitamins`,
    `MECPTP Maternal Glycan`,
    `MECPTP Maternal Nucleotide`,
    `MECPTP Maternal Energy`,
    `MECPTP Maternal Xenobiotics`,
    `MECPTP Maternal Other_secondary_metabolites`,
    #11
    `MEHHTP Maternal Carbohydrate`,
    `MEHHTP Maternal amino acid`,
    `MEHHTP Maternal Lipid`,
    `MEHHTP Maternal Cofactors_vitamins`,
    `MEHHTP Maternal Glycan`,
    `MEHHTP Maternal Nucleotide`,
    `MEHHTP Maternal Energy`,
    `MEHHTP Maternal Xenobiotics`,
    `MEHHTP Maternal Other_secondary_metabolites`
)
```

#overlap plots
```{r}
upset(dat_olap_crazy, 
      keep.order = TRUE,
      nsets = 20,
      sets.bar.color = c(
    rep("#0072B2", 9),       # Blue
    rep("#E69F00", 10),      # Orange
    rep("#009E73", 6),       # Green
    rep("#56B4E9", 9),       # Sky Blue
    rep("#FFD700", 9),       # Gold/Yellow
    rep("#008080", 7),       # Teal
    rep("#FFA500", 8),        # Light Orange
    rep("#AFA600", 9),        # Light Orange
    rep("#CFA600", 9)        # Light Orange
)
,
      sets = c("2,5-Dichlorophenol Newborn Carbohydrate",
"2,5-Dichlorophenol Newborn amino acid",
"2,5-Dichlorophenol Newborn Lipid",
"2,5-Dichlorophenol Newborn Cofactors_vitamins",
"2,5-Dichlorophenol Newborn Glycan",
"2,5-Dichlorophenol Newborn Nucleotide",
"2,5-Dichlorophenol Newborn Energy",
"2,5-Dichlorophenol Newborn Xenobiotics",
"2,5-Dichlorophenol Newborn Other_secondary_metabolites",
# 10
"BPA Newborn Carbohydrate",
"BPA Newborn amino acid",
"BPA Newborn Lipid",
"BPA Newborn Cofactors_vitamins",
"BPA Newborn Glycan",
"BPA Newborn Nucleotide",
"BPA Newborn Energy",
"BPA Newborn Xenobiotics",
"BPA Newborn Other_secondary_metabolites",
"BPA Newborn Terpenoids_polyketides",
# 6
"MBP Newborn Carbohydrate",
"MBP Newborn amino acid",
"MBP Newborn Lipid",
"MBP Newborn Cofactors_vitamins",
"MBP Newborn Xenobiotics",
"MBP Newborn Other_secondary_metabolites",
# 9
"MBzP Newborn Carbohydrate",
"MBzP Newborn amino acid",
"MBzP Newborn Lipid",
"MBzP Newborn Cofactors_vitamins",
"MBzP Newborn Glycan",
"MBzP Newborn Nucleotide",
"MBzP Newborn Energy",
"MBzP Newborn Xenobiotics",
"MBzP Newborn Other_secondary_metabolites",
# 9
"BPA Maternal Carbohydrate",
"BPA Maternal amino acid",
"BPA Maternal Lipid",
"BPA Maternal Cofactors_vitamins",
"BPA Maternal Glycan",
"BPA Maternal Nucleotide",
"BPA Maternal Energy",
"BPA Maternal Xenobiotics",
"BPA Maternal Other_secondary_metabolites",
# 7
"BPS Maternal Carbohydrate",
"BPS Maternal amino acid",
"BPS Maternal Lipid",
"BPS Maternal Energy",
"BPS Maternal Nucleotide",
"BPS Maternal Cofactors_vitamins",
"BPS Maternal Other_secondary_metabolites",
# 8
"MiBP Maternal Carbohydrate",
"MiBP Maternal amino acid",
"MiBP Maternal Lipid",
"MiBP Maternal Cofactors_vitamins",
"MiBP Maternal Glycan",
"MiBP Maternal Nucleotide",
"MiBP Maternal Xenobiotics",
"MiBP Maternal Other_secondary_metabolites",
#9
"MECPTP Maternal Carbohydrate",
"MECPTP Maternal amino acid",
"MECPTP Maternal Lipid",
"MECPTP Maternal Cofactors_vitamins",
"MECPTP Maternal Glycan",
"MECPTP Maternal Nucleotide",
"MECPTP Maternal Energy",
"MECPTP Maternal Xenobiotics",
"MECPTP Maternal Other_secondary_metabolites",
#9
"MEHHTP Maternal Carbohydrate",
"MEHHTP Maternal amino acid",
"MEHHTP Maternal Lipid",
"MEHHTP Maternal Cofactors_vitamins",
"MEHHTP Maternal Glycan",
"MEHHTP Maternal Nucleotide",
"MEHHTP Maternal Energy",
"MEHHTP Maternal Xenobiotics",
"MEHHTP Maternal Other_secondary_metabolites"
),
      set_size.show = TRUE,
      text.scale = 1.3, point.size = 3, line.size = 1,
      mainbar.y.max = 3,
      set_size.scale_max = 20,
      mainbar.y.label="No.of pathways \noverlapped across chemical biomarkers",
      sets.x.label="Total no.of \naltered pathways")
```

#overlapping pathwats content for text
```{r}
dat_olap_tbl<- dat |> select(c(1,12,13,17)) |> unite(filename, chemical, mom_baby, sep = "_") |>
  group_by(filename, group) |>
  summarise(pathway = paste(pathway, collapse = ', '), count = n()) |>
  ungroup()

filename_values_of_interest <- c("2,5-Dichlorophenol Newborn", "BPA Newborn")
filename_values_of_interest<- c("MiBP Maternal", "MEHHTP Maternal")
filename_values_of_interest <- c("BPA Maternal", "BPA Newborn")
filename_values_of_interest <- c("BPS Maternal", "BPS Newborn")

x_tab <- dat |>
  unite(filename, chemical, mom_baby, sep = " ") |>
  select(pathway, filename, group) |>
  filter(filename %in% filename_values_of_interest) |>
  group_by(group, pathway) |> #
  summarise(filename_count = n()) |>
  filter(filename_count == length(filename_values_of_interest)) |>
  group_by(group) |>
  summarise(total_common_pathways = n(),
            common_pathways = paste(pathway, collapse = ", "))

sum(x_tab$total_common_pathways)

# Print the total common pathways
cat("Total Common Pathways: ", sum(x_tab$total_common_pathways), "\n\n")

# Print the result for each group
for (grp in x_tab$group) {
  cat("Group Name: ", grp, "\n")
  cat("Total Common Pathways: ", x_tab$total_common_pathways[x_tab$group == grp], "\n")
  cat("Common Pathways:\n", x_tab$common_pathways[x_tab$group == grp], "\n\n")
}
```

#bettern bar plot
```{r}
# Calculate the total count for each variable
variable_counts <- colSums(dat_olap_crazy, na.rm = TRUE)

# Create a dataframe for the variable counts
variable_counts_df <- data.frame(variable = names(variable_counts), count = variable_counts)

ggplot(variable_counts_df, aes(x = count, y = variable, fill = substr(variable, 1, 4))) +
  geom_col() +
  geom_text(aes(label = count), position = position_dodge(width = 0.9), hjust = -0.1) +
  labs(title = " ", x = "Frequency of Enriched pathways", y = " ") +
  theme(legend.position = "none", axis.text.y = element_text(hjust = 1)) +
  guides(fill="none")+
  scale_y_discrete(expand = expansion(add = c(0, 0))) +
  scale_x_continuous(expand = expansion(mult = c(0, 0.05)))+
  theme_minimal()

ggsave("~/Documents/phth_phe_MWAS/result/pathway_histogram.tiff", 
       width = 10, height = 16, dpi=300)
```
