==> 88_kegg_super_pathways.R <==
pathways = list(
  `Carbohydrate` = c(
    "Glycolysis", "Gluconeogenesis", "Citrate cycle", "TCA cycle", "Pentose phosphate", "Pentose", "glucuronate",
    "Fructose", "mannose", "Galactose", "Ascorbate", "aldarate", "Starch", "sucrose",
    "Amino sugar", "nucleotide sugar", "Pyruvate", "Glyoxylate","dicarboxylate", "Propanoate ",
    "Butanoate", "C5-Branched", "Inositol phosphate", "Lipoate"
  ),
  `Energy` = c(
    "Oxidative phosphorylation", "Photosynthesis", "Photosynthesis - antenna proteins", "Carbon fixation",
    "Methane ", "Nitrogen", "Sulfur"
  ),
  `Lipid` = c(
    "Fatty acid", "Fatty acid elongation", "Cutin, suberine and wax",
    "Steroid", "Primary bile acid", "Secondary bile acid", "Steroid hormone",
    "Glycerolipid", "Glycerophospholipid", "Ether lipid", "Sphingolipid",
    "Arachidonic acid", "Linoleic acid", "alpha-Linolenic acid", "unsaturated fatty acids",
    "Leukotriene", "Carnitine shuttle", "Leukotriene", "Prostaglandin formation from arachidonate",
    "Linoleate", "Bile acid biosynthesis"
  ),
  `Nucleotide` = c(
    "Purine", "Pyrimidine"
  ),
  `amino acid` = c(
    "Alanine", "aspartate", "glutamate", "Glycine", "threonine",
    "Cysteine",  "methionine", "Valine", "leucine", "isoleucine",
    "Valine", "leucine", "isoleucine", "Lysine",
    "Lysine", "Arginine", "Arginine", "proline",
    "Histidine", "Tyrosine", "Phenylalanine",
    "Tryptophan", "Phenylalanine", "tyrosine", "tryptophan",
    "beta-Alanine", "Taurine", "hypotaurine",
    "Phosphonate", "phosphinate", "Selenocompound",
    "Cyanoamino", "D-Amino", "Glutathione", "Selenoamino", 
    "Sialic acid"
  ),
  `Glycan` = c(
    "N-Glycan", "O-glycan", "Glycosaminoglycan",
    "Glycosphingolipid",
    "Lipopolysaccharide", "O-Antigen", "Peptidoglycan", "Teichoic",
    "Lipoarabinomannan", "Arabinogalactan", "Exopolysaccharide ",
    "Blood Group Biosynthesis", "Glycosylphosphatidylinositol(GPI)-anchor biosynthesis",
    "Chondroitin sulfate degradation", "Heparan sulfate degradation",
    "Keratan sulfate degradation", "Keratan sulfate biosynthesis"
  ),
  `Cofactors_vitamins` = c(
    "Thiamine", "Riboflavin", "Vitamin B6",
    "Nicotinate and nicotinamide", "Pantothenate", "CoA",
    "Biotin", "Lipoic", "Folate",
    "One carbon pool by folate", "Retinol ", "Porphyrin",
    "Ubiquinone", "terpenoid-quinone ", "vitamin", "Biopterin"
  )
  ,
  `Terpenoids_polyketides` = c(
    "Terpenoid backbone", "Monoterpenoid",
    "Sesquiterpenoid", "triterpenoid", "Diterpenoid",
    "Gibberellin", "Carotenoid",
    "Brassinosteroid", "Insect hormone",
    "Zeatin", "Limonene",
    "Pinene", "camphor", "geraniol", "Type I polyketide structures",
    "12-, 14- and 16-membered macrolides",
    "Tylosin", "Mycinamicin",
    "Erythromycin", "Oleandomycin",
    "Pikromycin/methymycin", "Avermectin",
    "ansamycins", " of enediyne antibiotics",
    "type II polyketide backbone",
    "type II polyketide products", "Tetracycline",
    "Polyketide sugar unit", "Nonribosomal peptide structures",
    "siderophore group nonribosomal peptides",
    "vancomycin group antibiotics"
  ),
  `Other_secondary_metabolites` = c(
    "Phenylpropanoid", "Stilbenoid, diarylheptanoid and gingerol",
    "Flavonoid", "Flavone", "flavonol", "Anthocyanin",
    "Isoflavonoid", " flavonoids", "alkaloid ",
    "Cucurbitacin", "Solanine and tomatine", "Ephedrine",
    "Capsaicin", "Acridone alkaloid", "Caffeine",
    "Betalain", "Glucosinolate", "Benzoxazinoid",
    "Penicillin and cephalosporin", "Carbapenem", "Monobactam",
    "Clavulanic acid", "Streptomycin", "Neomycin, kanamycin and gentamicin",
    "Acarbose and validamycin", "Novobiocin", "Staurosporine",
    "Phenazine", "Prodigiosin", "Aflatoxin",
    "various antibiotics", "Kanosamine", "Aurachin",
    "Bacilysin", "Puromycin", "Dapdiamides",
    "Fosfomycin", "Cremeomycin", "Fumagillin",
    "Pentalenolactone", "Terpentecin", "Roseoflavin",
    "Cycloserine", "various plant secondary metabolites",
    "Crocin", "Ginsenoside", "Saponin",
    "Cannabidiol", "Mugineic acid", "Pentagalloylglucose",
    "Benzoxazinoid", "Gramine", "Coumarin",
    "Furanocoumarin", "Hordatin", "Podophyllotoxin",
    "other secondary metabolites", "Ditryptophenaline",
    "Fumiquinazoline D", "Paerucumarin", "Staphyloferrin B",
    "Cyclooctatin", "Lovastatin", "Grixazone",
    "Staphyloferrin A", "Ethynylserine", "Aerobactin"
  ),
  `Xenobiotics` = c(
    "Benzoate", "Aminobenzoate", "Fluorobenzoate",
    "Chloroalkane and chloroalkene", "Chlorocyclohexane", "chlorobenzene",
    "Toluene", "Xylene", "Nitrotoluene",
    "Ethylbenzene", "Styrene", "Atrazine", "Caprolactam",
    "Bisphenol", "Dioxin", "Naphthalene",
    "Polycyclic aromatic hydrocarbon", "Furfural", "Steroid",
    "xenobiotics", "Drug", "Benzoate degradation via CoA ligation",
    "Parathio degradation"
  ),
  `Chemical_structure_transformation_maps` = c(
    "plant secondary metabolites", "phenylpropanoids",
    "terpenoids and steroids", "alkaloids derived from shikimate pathway",
    "alkaloids derived from ornithine, nicotinic acid",
    "alkaloids derived from histidine and purine", "alkaloids derived from terpenoid and polyketide",
    "plant hormones"
  ),
  `NA` = c("Heparan sulfate degradation", "Keratan sulfate degradation", "Prostaglandin formation from arachidonate",
           "Putative anti-Inflammatory metabolites formation from EPA", "Hexose phosphorylation",
           "Phytanic acid peroxisomal oxidation", "D4&E4-neuroprostanes formation",
           "Hyaluronan Metabolism", "Electron transport chain")
)
